feat: build category and study material repository
