+++ src/upload-file/Category.ts
@@ -1,0 +1,4 @@
+export interface Category {
+    category: string;
+}
+
