+++ src/upload-file/StudyMaterial.ts
@@ -1,0 +1,9 @@
+export interface StudyMaterial{
+    filename: string;
+    filePath: string;
+    category: string;
+    title: string;
+    description: string;
+    date: Date;
+
+}
